+++ CHANGES.md
@@ -10,10 +10,9 @@ and this project adheres to
 ### Added
 
 - Implement pipelined reports (`TestRun.Report`, `Test.During`, `Test.Report`)
-- Turn Analyze into a report and add to examples that need it
+- Turn Analyze into a report and add it to examples that need it
 - Implement log sorting
-- Write results to single directory, configurable from CUE
-- Allow writing test results non-destructively (`Results.Destructive=false`)
+- Write test results non-destructively
 - Validate ResultPrefixes are unique
 - Add embedded web server to serve results (`server` command)